dplsh: CTRL-C issue
NOTE: I don't know if the committed code is a canonical way to deal with ^C in readline,
and we have a slight latency between the time when ^C is displayed and the new clean prompt.

From the readline documentation:
"Variable: rl_hook_func_t * rl_event_hook -- If non-zero, this is the address of a function
to call periodically when Readline is waiting for terminal input. By default, this
will be called at most ten times a second if there is no keyboard input."