droplet-sh: license problem
This tool is now under BSD license.

As a consequence, a libreadline facility is now disabled (the Ctrl-C behaviour),
because some callbacks/variables are not available with libedit-dev.

Note also that the editline headers are wrong about the entry name completion callback
prototype, so we do an ugly and explicit cast to the appropriate type.